update script for data preparation
